Fixed all tests runs, Build Success

+++ src/main/java/com/vut/calculator/GradeCalculator.java
@@ -90,7 +90,7 @@ public class GradeCalculator {
                 passCount++;
             }
         }
-        return passCount / finalMarks.length;
+        return (double )passCount / finalMarks.length;
     }
 
     /**
